feat(web): consolidate org-scoped settings experience
Unify settings into a single org-scoped section with nested pages and legacy redirects so users can access authentication, integrations, connections, team, appearance, and about from one extensible surface.
