+++ apps/web/src/routes/settings.test.tsx
@@ -18,47 +18,8 @@ const mocks = vi.hoisted(() => ({
     defaultPriority: number | null
     lastValidatedAt: string | null
   },
-  listAccounts: vi.fn(),
   saveLinearIntegrationMutateAsync: vi.fn(),
   testLinearIntegrationMutateAsync: vi.fn(),
-  trackEvent: vi.fn(),
-}))
-
-vi.mock('@tanstack/react-router', () => ({
-  createFileRoute: () => (options: unknown) => ({
-    options,
-  }),
-}))
-
-vi.mock('@durabull/analytics/browser', () => ({
-  trackEvent: mocks.trackEvent,
-}))
-
-vi.mock('@durabull/analytics/events', () => ({
-  AnalyticsEvents: {
-    DIALOG_CLOSED: 'DIALOG_CLOSED',
-    DIALOG_OPENED: 'DIALOG_OPENED',
-    USER_ACCOUNT_LINKED: 'USER_ACCOUNT_LINKED',
-    USER_ACCOUNT_UNLINKED: 'USER_ACCOUNT_UNLINKED',
-  },
-  DialogType: {
-    UNLINK_ACCOUNT: 'UNLINK_ACCOUNT',
-  },
-}))
-
-vi.mock('@/components/app-top-bar', () => ({
-  useAppTopBar: vi.fn(),
-}))
-
-vi.mock('@/hooks/use-app-config', () => ({
-  useAppConfig: () => ({
-    config: {
-      telemetry: {
-        collectionRequired: true,
-        disclosureUrl: 'https://durabull.io/privacy',
-      },
-    },
-  }),
 }))
 
 vi.mock('@/hooks/use-alerts', () => ({
@@ -81,58 +42,23 @@ vi.mock('@/hooks/use-alerts', () => ({
   }),
 }))
 
-vi.mock('@/hooks/use-auth', () => ({
-  linkSocial: vi.fn(),
-  listAccounts: mocks.listAccounts,
-  unlinkAccount: vi.fn(),
-  useAuth: () => ({
-    user: {
-      id: 'user-1',
-      email: 'test@example.com',
-    },
-    isLoading: false,
-  }),
-}))
+import { IntegrationsSettingsPanel } from '@/components/settings/integrations-settings-panel'
 
-vi.mock('@/lib/app-version', () => ({
-  APP_BUILD_INFO: {
-    version: '1.2.3-test',
-    buildId: 'test-build',
-    buildTime: null,
-  },
-}))
-
-import { Route } from '@/routes/settings'
-
-describe('SettingsPage', () => {
+describe('IntegrationsSettingsPanel', () => {
   beforeEach(() => {
-    window.sessionStorage.clear()
     mocks.connectLinearIntegrationMutateAsync.mockReset()
     mocks.connectLinearIntegrationMutateAsync.mockResolvedValue({
       authorizationUrl: 'https://linear.app/oauth/authorize?state=test',
     })
     mocks.deleteLinearIntegrationMutateAsync.mockReset()
     mocks.linearIntegration = null
-    mocks.listAccounts.mockReset()
-    mocks.listAccounts.mockResolvedValue({ data: [] })
     mocks.saveLinearIntegrationMutateAsync.mockReset()
     mocks.testLinearIntegrationMutateAsync.mockReset()
-    mocks.trackEvent.mockReset()
   })
 
-  it('shows the current app version quietly on the settings page', () => {
-    const Component = Route.options.component as () => React.ReactNode
-
-    render(<Component />)
-
-    expect(screen.getByText('Durabull v1.2.3-test')).toBeInTheDocument()
-  })
-
-  it('only shows the Linear connect action before OAuth is configured', async () => {
-    const Component = Route.options.component as () => React.ReactNode
+  it('shows the Linear connect action before OAuth is configured', async () => {
     mocks.connectLinearIntegrationMutateAsync.mockImplementation(() => new Promise(() => {}))
-
-    render(<Component />)
+    render(<IntegrationsSettingsPanel />)
 
     expect(
       screen.queryByRole('textbox', { name: /default linear team id/i })
@@ -163,9 +89,8 @@ describe('SettingsPage', () => {
     mocks.saveLinearIntegrationMutateAsync.mockResolvedValue({
       integration: mocks.linearIntegration,
     })
-    const Component = Route.options.component as () => React.ReactNode
 
-    render(<Component />)
+    render(<IntegrationsSettingsPanel />)
 
     const teamInput = screen.getByRole('textbox', { name: /default linear team id/i })
     expect(teamInput).toHaveValue('')
